Add citations to the code in problems 2 and 3.
Arguably, such citations are not necessary, since it's my code and
it's implicit that I reuse code from the current and previous course
offerings, including from labs and assignments.  However, I've added
the citations to demonstrate good practice.  (It's not great practice -
I haven't given dates of copying and such, but it suffices.)